@@ -1,12 +1,11 @@
-import math
-from typing import Any, Dict, Optional, Sequence, Union
-from uuid import UUID
+import logging
+from typing import AsyncIterator, Optional, Sequence, Union
 
-from anyio import create_memory_object_stream
-from langchain.callbacks.base import BaseCallbackHandler
-from langchain.schema.messages import (
+import orjson
+from langchain_core.messages import (
     AIMessage,
     AIMessageChunk,
+    AnyMessage,
     BaseMessage,
     BaseMessageChunk,
     ChatMessage,
@@ -16,54 +15,49 @@ from langchain.schema.messages import (
     HumanMessage,
     HumanMessageChunk,
 )
-from langchain.schema.output import ChatGenerationChunk, GenerationChunk
+from langchain_core.runnables import Runnable, RunnableConfig
+from langserve.serialization import WellKnownLCSerializer
+
+logger = logging.getLogger(__name__)
+
+MessagesStream = AsyncIterator[Union[list[AnyMessage], str]]
 
 
-class StreamMessagesHandler(BaseCallbackHandler):
-    def __init__(self, messages: Sequence[BaseMessage]) -> None:
-        self.messages = messages
-        self.output: Dict[UUID, ChatGenerationChunk] = {}
-        send_stream, receive_stream = create_memory_object_stream(math.inf)
-        self.send_stream = send_stream
-        self.receive_stream = receive_stream
+async def astream_messages(
+    app: Runnable, input: Sequence[AnyMessage], config: RunnableConfig
+) -> MessagesStream:
+    """Stream messages from the runnable."""
+    root_run_id: Optional[str] = None
+    last_messages_list: Optional[list[AnyMessage]] = None
+    last_stream_run_id: Optional[str] = None
 
-    def on_llm_new_token(
-        self,
-        token: str,
-        *,
-        chunk: Optional[Union[GenerationChunk, ChatGenerationChunk]] = None,
-        run_id: UUID,
-        **kwargs: Any,
-    ) -> Any:
-        # If this is being called for a non-Chat Model run, convert to AIMessage
-        if chunk is None:
-            chunk = ChatGenerationChunk(message=AIMessageChunk(content=token))
-        # If we get something we don't know how to handle, ignore it
-        if not (
-            isinstance(chunk, ChatGenerationChunk)
-            or isinstance(chunk, BaseMessageChunk)
+    async for event in app.astream_events(
+        input, config, version="v1", output_keys=["__root__"]
+    ):
+        if event["event"] == "on_chain_start" and not root_run_id:
+            root_run_id = event["run_id"]
+
+            yield root_run_id
+        elif event["event"] == "on_chain_stream" and event["run_id"] == root_run_id:
+            last_messages_list = event["data"]["chunk"]["__root__"]
+
+            yield last_messages_list
+        elif (
+            event["event"] == "on_chat_model_stream" and last_messages_list is not None
         ):
-            return
-        # Convert messages to ChatGenerationChunks (workaround for old langchahin)
-        if isinstance(chunk, BaseMessageChunk):
-            chunk = ChatGenerationChunk(message=chunk)
-        # Accumulate the output (ChatGenerationChunk implements __add__)
-        if not self.output.get(run_id):
-            self.output[run_id] = chunk
-        else:
-            self.output[run_id] += chunk
-        # Send the messages to the stream
-        self.send_stream.send_nowait(
-            {
-                "messages": (
-                    self.messages
-                    + [
-                        map_chunk_to_msg(chunk.message)
-                        for chunk in self.output.values()
-                    ]
-                )
-            }
-        )
+            is_new_stream_run = (
+                last_stream_run_id is None or last_stream_run_id != event["run_id"]
+            )
+            is_diff_msg_type = last_messages_list and type(  # noqa: E721
+                last_messages_list[-1]
+            ) != type(event["data"]["chunk"])
+            if is_new_stream_run or is_diff_msg_type:
+                last_stream_run_id = event["run_id"]
+                last_messages_list.append(event["data"]["chunk"])
+            else:
+                last_messages_list[-1] = last_messages_list[-1] + event["data"]["chunk"]
+
+            yield last_messages_list
 
 
 def map_chunk_to_msg(chunk: BaseMessageChunk) -> BaseMessage:
@@ -80,3 +74,41 @@ def map_chunk_to_msg(chunk: BaseMessageChunk) -> BaseMessage:
         return ChatMessage(**args)
     else:
         raise ValueError(f"Unknown chunk type: {chunk}")
+
+
+_serializer = WellKnownLCSerializer()
+
+
+async def to_sse(messages_stream: MessagesStream) -> AsyncIterator[dict]:
+    """Consume the stream into an EventSourceResponse"""
+    try:
+        async for chunk in messages_stream:
+            # EventSourceResponse expects a string for data
+            # so after serializing into bytes, we decode into utf-8
+            # to get a string.
+            if isinstance(chunk, str):
+                yield {
+                    "event": "metadata",
+                    "data": orjson.dumps({"run_id": chunk}).decode(),
+                }
+            else:
+                yield {
+                    "event": "data",
+                    "data": _serializer.dumps(
+                        [map_chunk_to_msg(msg) for msg in chunk]
+                    ).decode(),
+                }
+    except Exception:
+        logger.warn("error in stream", exc_info=True)
+        yield {
+            "event": "error",
+            # Do not expose the error message to the client since
+            # the message may contain sensitive information.
+            # We'll add client side errors for validation as well.
+            "data": orjson.dumps(
+                {"status_code": 500, "message": "Internal Server Error"}
+            ).decode(),
+        }
+
+    # Send an end event to signal the end of the stream
+    yield {"event": "end"}
