@@ -8,11 +8,11 @@ know about server/uploading etc.
 """
 from typing import List
 
-from langchain.document_loaders import Blob
-from langchain.document_loaders.base import BaseBlobParser
-from langchain.schema import Document
-from langchain.schema.vectorstore import VectorStore
 from langchain.text_splitter import TextSplitter
+from langchain_community.document_loaders import Blob
+from langchain_community.document_loaders.base import BaseBlobParser
+from langchain_core.documents import Document
+from langchain_core.vectorstores import VectorStore
 
 
 def _update_document_metadata(document: Document, namespace: str) -> None:
