@@ -8,25 +8,32 @@ For the time being, upload and ingestion are coupled
 """
 from __future__ import annotations
 
+import os
 from typing import Any, BinaryIO, List, Optional
 
-from langchain.document_loaders.blob_loaders.schema import Blob
-from langchain.schema.runnable import RunnableConfig, RunnableSerializable
-from langchain.schema.vectorstore import VectorStore
-from langchain.text_splitter import TextSplitter
+from langchain.text_splitter import RecursiveCharacterTextSplitter, TextSplitter
+from langchain_community.document_loaders.blob_loaders import Blob
+from langchain_community.vectorstores.redis import Redis
+from langchain_core.runnables import (
+    ConfigurableField,
+    RunnableConfig,
+    RunnableSerializable,
+)
+from langchain_core.vectorstores import VectorStore
+from langchain_openai import OpenAIEmbeddings
 
-from agent_executor.ingest import ingest_blob
-from agent_executor.parsing import MIMETYPE_BASED_PARSER
+from app.ingest import ingest_blob
+from app.parsing import MIMETYPE_BASED_PARSER
 
 
 def _guess_mimetype(file_bytes: bytes) -> str:
     """Guess the mime-type of a file."""
     try:
         import magic
-    except ImportError:
+    except ImportError as e:
         raise ImportError(
             "magic package not found, please install it with `pip install python-magic`"
-        )
+        ) from e
 
     mime = magic.Magic(mime=True)
     mime_type = mime.from_buffer(file_bytes)
@@ -94,3 +101,26 @@ class IngestRunnable(RunnableSerializable[BinaryIO, List[str]]):
                 )
             )
         return ids
+
+
+index_schema = {
+    "tag": [{"name": "namespace"}],
+}
+vstore = Redis(
+    redis_url=os.environ["REDIS_URL"],
+    index_name="opengpts",
+    embedding=OpenAIEmbeddings(),
+    index_schema=index_schema,
+)
+
+
+ingest_runnable = IngestRunnable(
+    text_splitter=RecursiveCharacterTextSplitter(chunk_size=1000, chunk_overlap=200),
+    vectorstore=vstore,
+).configurable_fields(
+    assistant_id=ConfigurableField(
+        id="assistant_id",
+        annotation=str,
+        name="Assistant ID",
+    ),
+)
