@@ -1,8 +1,8 @@
 from io import BytesIO
 
-from agent_executor.upload import IngestRunnable, _guess_mimetype
 from langchain.text_splitter import RecursiveCharacterTextSplitter
 
+from app.upload import IngestRunnable, _guess_mimetype
 from tests.unit_tests.fixtures import get_sample_paths
 from tests.unit_tests.utils import InMemoryVectorStore
 
