@@ -1,15 +1,69 @@
-import { useEffect, useState } from "react";
-import { marked } from "marked";
 import { ShareIcon } from "@heroicons/react/24/outline";
+import { last, orderBy } from "lodash";
+import { marked } from "marked";
+import { useEffect, useState } from "react";
 import { useDropzone } from "react-dropzone";
-import { orderBy } from "lodash";
 
+import { Switch } from "@headlessui/react";
+import { TYPES } from "../constants";
 import { ConfigListProps } from "../hooks/useConfigList";
 import { SchemaField, Schemas } from "../hooks/useSchemas";
 import { cn } from "../utils/cn";
 import { FileUploadDropzone } from "./file-upload/FileUpload";
 
-function Label(props: { id: string; title: string }) {
+function Types(props: {
+  field: SchemaField;
+  value: string;
+  readonly: boolean;
+  setValue: (value: string) => void;
+  alwaysExpanded?: boolean;
+}) {
+  const options =
+    props.field.enum?.map((id) => TYPES[id as keyof typeof TYPES]) ?? [];
+  return (
+    <div className="-mx-8 mt-6 pt-4 border-t-2 border-dotted mb-8">
+      <div className="mx-8 sm:hidden">
+        <label htmlFor="tabs" className="sr-only">
+          Select a tab
+        </label>
+        <select
+          id="tabs"
+          name="tabs"
+          className="block w-full rounded-md border-gray-300 focus:border-indigo-500 focus:ring-indigo-500"
+          defaultValue={options.find((o) => o.id === props.value)?.id}
+          onChange={(e) => props.setValue(e.target.value)}
+        >
+          {options.map((option) => (
+            <option key={option.id}>{option.title}</option>
+          ))}
+        </select>
+      </div>
+      <div className="mx-8 hidden sm:block">
+        <div className="border-b border-gray-200">
+          <nav className="-mb-px flex" aria-label="Tabs">
+            {options.map((option) => (
+              <div
+                key={option.id}
+                className={cn(
+                  props.value === option.id
+                    ? "border-indigo-500 text-indigo-600"
+                    : "border-transparent text-gray-500 hover:border-gray-300 hover:text-gray-700",
+                  "w-1/4 border-b-2 py-4 px-1 text-center text-sm font-medium cursor-pointer"
+                )}
+                aria-current={props.value === option.id ? "page" : undefined}
+                onClick={() => props.setValue(option.id)}
+              >
+                {option.title}
+              </div>
+            ))}
+          </nav>
+        </div>
+      </div>
+    </div>
+  );
+}
+
+function Label(props: { id?: string; title: string }) {
   return (
     <label
       htmlFor={props.id}
@@ -55,7 +109,7 @@ export default function SingleOptionField(props: {
 }) {
   return (
     <div>
-      <Label id={props.id} title={props.title} />
+      <Label id={props.id} title={props.field.title} />
       <fieldset>
         <legend className="sr-only">{props.field.title}</legend>
         <div className="space-y-2">
@@ -160,7 +214,10 @@ function MultiOptionField(props: {
 }
 
 function PublicLink(props: { assistantId: string }) {
-  const link = window.location.href + "?shared_id=" + props.assistantId;
+  const currentLink = window.location.href;
+  const link = currentLink.includes("shared_id=")
+    ? currentLink
+    : currentLink + "?shared_id=" + props.assistantId;
   return (
     <div className="flex rounded-md shadow-sm mb-4">
       <button
@@ -189,7 +246,50 @@ function PublicLink(props: { assistantId: string }) {
   );
 }
 
+function PublicToggle(props: {
+  enabled: boolean;
+  setEnabled: (enabled: boolean) => void;
+  className?: string;
+}) {
+  return (
+    <Switch.Group as="div" className={cn("flex items-center", props.className)}>
+      <Switch
+        checked={props.enabled}
+        onChange={props.setEnabled}
+        className={cn(
+          props.enabled ? "bg-indigo-600" : "bg-gray-200",
+          "relative inline-flex h-6 w-11 flex-shrink-0 cursor-pointer rounded-full border-2 border-transparent transition-colors duration-200 ease-in-out focus:outline-none focus:ring-2 focus:ring-indigo-600 focus:ring-offset-2"
+        )}
+      >
+        <span
+          aria-hidden="true"
+          className={cn(
+            props.enabled ? "translate-x-5" : "translate-x-0",
+            "pointer-events-none inline-block h-5 w-5 transform rounded-full bg-white shadow ring-0 transition duration-200 ease-in-out"
+          )}
+        />
+      </Switch>
+      <Switch.Label as="span" className="ml-3 text-sm">
+        <span className="font-medium text-gray-900">Public?</span>
+      </Switch.Label>
+    </Switch.Group>
+  );
+}
+
+function fileId(file: File) {
+  return `${file.name}-${file.size}-${file.lastModified}`;
+}
+
+const ORDER = [
+  "system_message",
+  "tools",
+  "llm_type",
+  "agent_type",
+  "retrieval_description",
+];
+
 export function Config(props: {
+  className?: string;
   configSchema: Schemas["configSchema"];
   configDefaults: Schemas["configDefaults"];
   config: ConfigListProps["currentConfig"];
@@ -237,7 +337,7 @@ export function Config(props: {
         };
       });
       setFiles((files) => [
-        ...files.filter((f) => !dropzone.acceptedFiles.includes(f)),
+        ...files.filter((f) => !acceptedFileIds.includes(fileId(f))),
         ...dropzone.acceptedFiles,
       ]);
     }
@@ -245,19 +345,66 @@ export function Config(props: {
 
   const [inflight, setInflight] = useState(false);
   const readonly = !!props.config && !inflight;
-  return (
-    <>
-      <div className="flex gap-2 items-center justify-between font-semibold text-lg leading-6 text-gray-600 mb-4">
-        <span>
-          Bot: {props.config?.name ?? "New Bot"}{" "}
-          <span className="font-normal">
-            {props.config ? "(read-only)" : ""}
-          </span>
-        </span>
+  const typeKey = "type";
+  const typeField =
+    props.configSchema?.properties.configurable.properties[typeKey];
+  const typeValue = values?.configurable?.[typeKey];
+  const typeSpec = typeValue ? TYPES[typeValue as keyof typeof TYPES] : null;
+
+  const settings = !props.config ? (
+    <div className="flex flex-row gap-4">
+      <div className="flex flex-row flex-1">
+        <div className="relative flex flex-grow items-stretch focus-within:z-10">
+          <input
+            type="text"
+            name="key"
+            id="key"
+            autoComplete="off"
+            className="block w-full rounded-none rounded-l-md border-0 py-1.5 pl-4 text-gray-900 ring-1 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6 ring-inset ring-gray-300"
+            placeholder="Name your bot"
+          />
+        </div>
+        <button
+          type="submit"
+          className="inline-flex items-center px-4 py-2 border border-transparent text-sm leading-5 font-medium rounded-r-md shadow-sm text-white bg-indigo-600 hover:bg-indigo-500 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-600"
+        >
+          {inflight ? "Saving..." : "Save"}
+        </button>
       </div>
-      {props.config?.public && (
+
+      <PublicToggle enabled={isPublic} setEnabled={setPublic} />
+    </div>
+  ) : (
+    <>
+      {props.config.public && (
         <PublicLink assistantId={props.config?.assistant_id} />
       )}
+    </>
+  );
+  return (
+    <div className={props.className}>
+      {settings}
+      {typeField && (
+        <Types
+          field={typeField}
+          value={typeValue as string}
+          setValue={(value: string) =>
+            setValues({
+              ...values,
+              configurable: { ...values!.configurable, [typeKey]: value },
+            })
+          }
+          readonly={readonly}
+        />
+      )}
+
+      {typeSpec?.description && (
+        <>
+          <Label title="Description" />
+          <div className="prose mb-8">{typeSpec.description}</div>
+        </>
+      )}
+
       <form
         className={cn("flex flex-col gap-8")}
         onSubmit={async (e) => {
@@ -268,12 +415,7 @@ export function Config(props: {
           if (!key) return;
 
           setInflight(true);
-          await props.saveConfig(
-            key,
-            values!,
-            dropzone.acceptedFiles,
-            isPublic
-          );
+          await props.saveConfig(key, values!, files, isPublic);
           setInflight(false);
         }}
       >
@@ -283,18 +425,20 @@ export function Config(props: {
             readonly && "opacity-50 cursor-not-allowed"
           )}
         >
-          {Object.entries(
-            props.configSchema?.properties.configurable.properties ?? {}
+          {orderBy(
+            Object.entries(
+              props.configSchema?.properties.configurable.properties ?? {}
+            ),
+            ([key]) => ORDER.indexOf(last(key.split("/"))!)
           ).map(([key, value]) => {
             const title = value.title;
-            if (value.allOf?.length === 1) {
-              value = value.allOf[0];
-            }
             if (key.split("/")[0].includes("==")) {
               const [parentKey, parentValue] = key.split("/")[0].split("==");
               if (values?.configurable?.[parentKey] !== parentValue) {
                 return null;
               }
+            } else {
+              return null;
             }
             if (value.type === "string" && value.enum) {
               return (
@@ -313,7 +457,7 @@ export function Config(props: {
                   readonly={readonly}
                 />
               );
-            } else if (key === "type==agent/system_message") {
+            } else if (value.type === "string") {
               return (
                 <StringField
                   key={key}
@@ -330,7 +474,11 @@ export function Config(props: {
                   readonly={readonly}
                 />
               );
-            } else if (key === "type==agent/tools") {
+            } else if (
+              value.type === "array" &&
+              value.items?.type === "string" &&
+              value.items?.enum
+            ) {
               return (
                 <MultiOptionField
                   key={key}
@@ -350,48 +498,15 @@ export function Config(props: {
               );
             }
           })}
-          {!props.config && values?.configurable?.type === "agent" && (
+          {!props.config && typeSpec?.files && (
             <FileUploadDropzone
               state={dropzone}
               files={files}
               setFiles={setFiles}
             />
           )}
-          <SingleOptionField
-            id="public"
-            field={{
-              type: "string",
-              title: "public",
-              description: "",
-              enum: ["Yes", "No"],
-            }}
-            title="Create a public link?"
-            value={isPublic ? "Yes" : "No"}
-            setValue={(value: string) => setPublic(value === "Yes")}
-            readonly={readonly}
-          />
         </div>
-        {!props.config && (
-          <div className="flex flex-row">
-            <div className="relative flex flex-grow items-stretch focus-within:z-10">
-              <input
-                type="text"
-                name="key"
-                id="key"
-                autoComplete="off"
-                className="block w-full rounded-none rounded-l-md border-0 py-1.5 pl-4 text-gray-900 ring-1 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6 ring-inset ring-gray-300"
-                placeholder="Name your bot"
-              />
-            </div>
-            <button
-              type="submit"
-              className="inline-flex items-center px-4 py-2 border border-transparent text-sm leading-5 font-medium rounded-r-md shadow-sm text-white bg-indigo-600 hover:bg-indigo-500 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-600"
-            >
-              {inflight ? "Saving..." : "Save"}
-            </button>
-          </div>
-        )}
       </form>
-    </>
+    </div>
   );
 }
