@@ -1,5 +1,4 @@
-import { PlusIcon } from "@heroicons/react/24/outline";
-
+import { TYPES } from "../constants";
 import { Config, ConfigListProps } from "../hooks/useConfigList";
 import { cn } from "../utils/cn";
 
@@ -16,7 +15,7 @@ function ConfigItem(props: {
           props.config === props.currentConfig
             ? "bg-gray-50 text-indigo-600"
             : "text-gray-700 hover:text-indigo-600 hover:bg-gray-50",
-          "group flex gap-x-3 rounded-md p-2 text-sm leading-6 font-semibold cursor-pointer"
+          "group flex gap-x-3 rounded-md p-2 leading-6 cursor-pointer"
         )}
       >
         <span
@@ -29,7 +28,19 @@ function ConfigItem(props: {
         >
           {props.config.name?.[0] ?? " "}
         </span>
-        <span className="truncate">{props.config.name}</span>
+        <div className="flex flex-col">
+          <span className="truncate text-sm font-medium">
+            {props.config.name}
+          </span>
+          <span className="truncate text-xs">
+            {
+              TYPES[
+                (props.config.config.configurable?.type ??
+                  "assistant") as keyof typeof TYPES
+              ].title
+            }
+          </span>
+        </div>
       </div>
     </li>
   );
@@ -42,29 +53,7 @@ export function ConfigList(props: {
 }) {
   return (
     <>
-      <div
-        onClick={() => props.enterConfig(null)}
-        className={cn(
-          props.currentConfig === null
-            ? "bg-gray-50 text-indigo-600"
-            : "text-gray-700 hover:text-indigo-600 hover:bg-gray-50",
-          "group flex gap-x-3 rounded-md -mx-2 p-2 text-sm leading-6 font-semibold cursor-pointer"
-        )}
-      >
-        <span
-          className={cn(
-            props.currentConfig === null
-              ? "text-indigo-600 border-indigo-600"
-              : "text-gray-400 border-gray-200 group-hover:border-indigo-600 group-hover:text-indigo-600",
-            "flex h-6 w-6 shrink-0 items-center justify-center rounded-lg border text-[0.625rem] font-medium bg-white"
-          )}
-        >
-          <PlusIcon className="h-4 w-4" />
-        </span>
-        <span className="truncate">New Bot</span>
-      </div>
-
-      <div className="text-xs font-semibold leading-6 text-gray-400 mt-4">
+      <div className="text-xs font-semibold leading-6 text-gray-400">
         Your Saved Bots
       </div>
       <ul role="list" className="-mx-2 mt-2 space-y-1">
