@@ -1,7 +1,7 @@
 import { ChevronDownIcon } from "@heroicons/react/24/outline";
 import DOMPurify from "dompurify";
 import { marked } from "marked";
-import { useState } from "react";
+import { memo, useState } from "react";
 import { Message as MessageType } from "../hooks/useChatList";
 import { cn } from "../utils/cn";
 import { str } from "../utils/str";
@@ -86,7 +86,9 @@ function Function(props: {
   );
 }
 
-export function Message(props: MessageType & { runId?: string }) {
+export const Message = memo(function Message(
+  props: MessageType & { runId?: string }
+) {
   const [open, setOpen] = useState(false);
   return (
     <div className="flex flex-col mb-8">
@@ -139,4 +141,4 @@ export function Message(props: MessageType & { runId?: string }) {
       )} */}
     </div>
   );
-}
+});
