@@ -12,7 +12,7 @@ export interface StreamState {
 export interface StreamStateProps {
   stream: StreamState | null;
   startStream: (
-    input: { messages: Message[] },
+    input: Message[],
     assistant_id: string,
     thread_id: string
   ) => Promise<void>;
@@ -24,14 +24,10 @@ export function useStreamState(): StreamStateProps {
   const [controller, setController] = useState<AbortController | null>(null);
 
   const startStream = useCallback(
-    async (
-      input: { messages: Message[] },
-      assistant_id: string,
-      thread_id: string
-    ) => {
+    async (input: Message[], assistant_id: string, thread_id: string) => {
       const controller = new AbortController();
       setController(controller);
-      setCurrent({ status: "inflight", messages: input.messages, merge: true });
+      setCurrent({ status: "inflight", messages: input, merge: true });
 
       await fetchEventSource("/runs/stream", {
         signal: controller.signal,
@@ -41,7 +37,7 @@ export function useStreamState(): StreamStateProps {
         openWhenHidden: true,
         onmessage(msg) {
           if (msg.event === "data") {
-            const { messages } = JSON.parse(msg.data);
+            const messages = JSON.parse(msg.data);
             setCurrent((current) => ({
               status: "inflight",
               messages,
@@ -53,6 +49,7 @@ export function useStreamState(): StreamStateProps {
               status: "inflight",
               messages: current?.messages,
               run_id: run_id,
+              merge: current?.merge,
             }));
           } else if (msg.event === "error") {
             setCurrent((current) => ({
